autoops/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Bootstrap
paramiko
APScheduler
requests
python-dotenv
PyMySQL
